@@ -1,44 +1,24 @@
-#!/usr/bin/env python3
-
-# Copyright 2022 by Michał Czuba, Piotr Bródka. All Rights Reserved.
-#
-# This file is part of Network Diffusion.
-#
-# Network Diffusion is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the Free
-# Software Foundation; either version 3 of the License, or (at your option) any
-# later version.
-#
-# Network Diffusion is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the  GNU General Public License for
-# more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# Network Diffusion. If not, see <http://www.gnu.org/licenses/>.
-# =============================================================================
-
 [build-system]
-requires = ["setuptools>=61", "setuptools-scm"]
+requires = ["setuptools~=69.5.1", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "network_diffusion"
-version = "0.14.4"
-requires-python = ">=3.10"
+version = "0.15.0"
+requires-python = ">=3.12"
 authors = [
     {name = "Michał Czuba", email="michal.czuba@pwr.edu.pl"},
     {name = "Piotr Bródka", email="piotr.brodka@pwr.edu.pl"},
 ]
-description = "Network Diffusion is a package for simulating spreading phenomena."
+description = "Network Diffusion - a package for simulating spreading phenomena."
 license = {file = "LICENSE"}
 classifiers = [
-    "Development Status :: 3 - Alpha",
+    "Development Status :: 4 - Beta",
     "Intended Audience :: Science/Research",
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
     "Topic :: Scientific/Engineering :: Information Analysis",
-    "License :: OSI Approved :: GNU General Public License (GPL)",
-    "Programming Language :: Python :: 3.10",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3.12",
 ]
 keywords = [
     "influence maximisation",
