@@ -1,30 +1,20 @@
-# Copyright 2023 by Michał Czuba, Piotr Bródka. All Rights Reserved.
+# Copyright (c) 2023 by Michał Czuba, Piotr Bródka.
 #
-# This file is part of Network Diffusion.
+# This file is a part of Network Diffusion.
 #
-# Network Diffusion is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the Free
-# Software Foundation; either version 3 of the License, or (at your option) any
-# later version.
-#
-# Network Diffusion is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the  GNU General Public License for
-# more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# Network Diffusion. If not, see <http://www.gnu.org/licenses/>.
+# Network Diffusion is licensed under the MIT License. You may obtain a copy
+# of the License at https://opensource.org/licenses/MIT
 # =============================================================================
 
 """Contains funcitons to handle actors of the multilayer network."""
 
-from typing import Any, Dict, Tuple
+from typing import Any
 
 
 class MLNetworkActor:
     """Dataclass that contain data of actor in the network."""
 
-    def __init__(self, actor_id: str, layers_states: Dict[str, str]) -> None:
+    def __init__(self, actor_id: str, layers_states: dict[str, str]) -> None:
         """
         Initialise the object.
 
@@ -36,7 +26,7 @@ class MLNetworkActor:
         self._layers_states = layers_states
 
     @classmethod
-    def from_dict(cls, base_dict: Dict[str, Any]) -> "MLNetworkActor":
+    def from_dict(cls, base_dict: dict[str, Any]) -> "MLNetworkActor":
         """
         Create an object from serialised dicitonary.
 
@@ -73,23 +63,23 @@ class MLNetworkActor:
         )
 
     @property
-    def layers(self) -> Tuple[str, ...]:
+    def layers(self) -> tuple[str, ...]:
         """Get network layers where actor exists."""
         return tuple(self._layers_states.keys())
 
     @property
-    def states(self) -> Dict[str, str]:
+    def states(self) -> dict[str, str]:
         """Get actor's states for where actitor exists."""
         return self._layers_states
 
     @states.setter
-    def states(self, updated_states: Dict[str, str]) -> None:
+    def states(self, updated_states: dict[str, str]) -> None:
         """Set actor's states for layers where it exists."""
         for layer_name, new_state in updated_states.items():
             assert layer_name in self._layers_states
             self._layers_states[layer_name] = new_state
 
-    def states_as_compartmental_graph(self) -> Tuple[str, ...]:
+    def states_as_compartmental_graph(self) -> tuple[str, ...]:
         """
         Return actor states in form accepted by CompartmentalGraph.
 
