@@ -1,26 +1,16 @@
-# Copyright 2023 by Michał Czuba, Piotr Bródka. All Rights Reserved.
+# Copyright (c) 2023 by Michał Czuba, Piotr Bródka.
 #
-# This file is part of Network Diffusion.
+# This file is a part of Network Diffusion.
 #
-# Network Diffusion is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the Free
-# Software Foundation; either version 3 of the License, or (at your option) any
-# later version.
-#
-# Network Diffusion is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the  GNU General Public License for
-# more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# Network Diffusion. If not, see <http://www.gnu.org/licenses/>.
+# Network Diffusion is licensed under the MIT License. You may obtain a copy
+# of the License at https://opensource.org/licenses/MIT
 # =============================================================================
 
 """Definition of the base propagation model used in the library."""
 
 from abc import ABC, abstractmethod
 from collections import Counter
-from typing import Any, Dict, List, Tuple
+from typing import Any
 
 from network_diffusion.mln.mlnetwork import MultilayerNetwork
 from network_diffusion.models.utils.compartmental import CompartmentalGraph
@@ -45,15 +35,10 @@ class BaseModel(ABC):
     def __str__(self) -> str:
         """Return string representation of the object."""
 
-    @property
-    def compartments(self) -> CompartmentalGraph:
-        """Return defined compartments and allowed transitions."""
-        return self._compartmental_graph
-
     @abstractmethod
     def determine_initial_states(
         self, net: MultilayerNetwork
-    ) -> List[NetworkUpdateBuffer]:
+    ) -> list[NetworkUpdateBuffer]:
         """
         Determine initial states in the network according to seed selector.
 
@@ -81,7 +66,7 @@ class BaseModel(ABC):
     @abstractmethod
     def network_evaluation_step(
         self, net: MultilayerNetwork
-    ) -> List[NetworkUpdateBuffer]:
+    ) -> list[NetworkUpdateBuffer]:
         """
         Evaluate the network at one time stamp according to the model.
 
@@ -92,8 +77,8 @@ class BaseModel(ABC):
     @staticmethod
     def update_network(
         net: MultilayerNetwork,
-        activated_nodes: List[NetworkUpdateBuffer],
-    ) -> List[Dict[str, str]]:
+        activated_nodes: list[NetworkUpdateBuffer],
+    ) -> list[dict[str, str]]:
         """
         Update the network global state by list of already activated nodes.
 
@@ -111,7 +96,7 @@ class BaseModel(ABC):
     @abstractmethod
     def get_allowed_states(
         self, net: MultilayerNetwork
-    ) -> Dict[str, Tuple[str, ...]]:
+    ) -> dict[str, tuple[str, ...]]:
         """
         Return dict with allowed states in each layer of net if applied model.
 
@@ -121,7 +106,7 @@ class BaseModel(ABC):
     @staticmethod
     def get_states_num(
         net: MultilayerNetwork,
-    ) -> Dict[str, Tuple[Tuple[Any, int], ...]]:
+    ) -> dict[str, tuple[tuple[Any, int], ...]]:
         """
         Return states in the network with number of agents that adopted them.
 
