@@ -1,6 +1,14 @@
+# Copyright (c) 2023 by Michał Czuba, Piotr Bródka.
+#
+# This file is a part of Network Diffusion.
+#
+# Network Diffusion is licensed under the MIT License. You may obtain a copy
+# of the License at https://opensource.org/licenses/MIT
+# =============================================================================
+
 """A definition of the seed selector based on betweenness centrality."""
 
-from typing import Any, List
+from typing import Any
 
 import networkx as nx
 
@@ -21,15 +29,15 @@ class BetweennessSelector(BaseSeedSelector):
             f"\tbetweenness centrality choice\n{BOLD_UNDERLINE}\n"
         )
 
-    def _calculate_ranking_list(self, graph: nx.Graph) -> List[Any]:
+    def _calculate_ranking_list(self, graph: nx.Graph) -> list[Any]:
         """Create nodewise ranking."""
         raise NotImplementedError(
             "Nodewise ranking list cannot be computed for this class!"
         )
 
-    def actorwise(self, net: MultilayerNetwork) -> List[MLNetworkActor]:
+    def actorwise(self, net: MultilayerNetwork) -> list[MLNetworkActor]:
         """Get ranking for actors using betweenness centrality metric."""
-        ranking_list: List[MLNetworkActor] = []
+        ranking_list: list[MLNetworkActor] = []
         # sorted by value of dictonaries
         for dc_val, _ in sorted(
             betweenness(net=net).items(), reverse=True, key=lambda x: x[1]
