@@ -1,24 +1,14 @@
-# Copyright 2023 by Michał Czuba, Piotr Bródka. All Rights Reserved.
+# Copyright (c) 2023 by Michał Czuba, Piotr Bródka.
 #
-# This file is part of Network Diffusion.
+# This file is a part of Network Diffusion.
 #
-# Network Diffusion is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the Free
-# Software Foundation; either version 3 of the License, or (at your option) any
-# later version.
-#
-# Network Diffusion is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the  GNU General Public License for
-# more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# Network Diffusion. If not, see <http://www.gnu.org/licenses/>.
+# Network Diffusion is licensed under the MIT License. You may obtain a copy
+# of the License at https://opensource.org/licenses/MIT
 # =============================================================================
 
 """A definition of "selector" that returns aprioiry provided actors."""
 
-from typing import Any, List
+from typing import Any
 
 import networkx as nx
 
@@ -31,7 +21,7 @@ from network_diffusion.utils import BOLD_UNDERLINE, THIN_UNDERLINE
 class MockingActorSelector(BaseSeedSelector):
     """Mock seed selector - returns a ranking provided as argument to init."""
 
-    def __init__(self, preselected_actors: List[MLNetworkActor]) -> None:
+    def __init__(self, preselected_actors: list[MLNetworkActor]) -> None:
         """Initialise object."""
         self.preselected_actors = preselected_actors
         super().__init__()
@@ -43,13 +33,13 @@ class MockingActorSelector(BaseSeedSelector):
             f"\tmocked choice - seeds provided a priori\n{BOLD_UNDERLINE}\n"
         )
 
-    def _calculate_ranking_list(self, graph: nx.Graph) -> List[Any]:
+    def _calculate_ranking_list(self, graph: nx.Graph) -> list[Any]:
         """Create nodewise ranking."""
         raise NotImplementedError(
             "Nodewise ranking list cannot be computed for this class!"
         )
 
-    def actorwise(self, net: MultilayerNetwork) -> List[MLNetworkActor]:
+    def actorwise(self, net: MultilayerNetwork) -> list[MLNetworkActor]:
         """Get ranking for actors."""
         if net.get_actors_num() != len(self.preselected_actors):
             raise ValueError("Incorrect length of preselected actors!")
