@@ -1,26 +1,14 @@
-#!/usr/bin/env python3
-
-# Copyright 2023 by Yuxuan Qiu. All Rights Reserved.
+# Copyright (c) 2023 by Yuxuan Qiu, Michał Czuba.
 #
-# This file is part of Network Diffusion.
+# This file is a part of Network Diffusion.
 #
-# Network Diffusion is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the Free
-# Software Foundation; either version 3 of the License, or (at your option) any
-# later version.
-#
-# Network Diffusion is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE. See the  GNU General Public License for
-# more details.
-#
-# You should have received a copy of the GNU General Public License along with
-# Network Diffusion. If not, see <http://www.gnu.org/licenses/>.
+# Network Diffusion is licensed under the MIT License. You may obtain a copy
+# of the License at https://opensource.org/licenses/MIT
 # =============================================================================
 
 """A script where a temporal network is defined."""
 
-from typing import Any, List, Optional
+from typing import Any
 
 import networkx as nx
 import pandas as pd
@@ -31,7 +19,7 @@ from network_diffusion.tpn.cogsnet_lib import _cogsnet  # pylint: disable=E0611
 from network_diffusion.utils import read_tpn
 
 
-def cogsnet_snap_to_nxgraph(cogsnet_snap: List[List[float]]) -> nx.DiGraph:
+def cogsnet_snap_to_nxgraph(cogsnet_snap: list[list[float]]) -> nx.DiGraph:
     """
     Create nx.DiGraph from one snapshot obtained from CogSNet.
 
@@ -49,7 +37,7 @@ def cogsnet_snap_to_nxgraph(cogsnet_snap: List[List[float]]) -> nx.DiGraph:
 class TemporalNetwork:
     """Container for a temporal network."""
 
-    def __init__(self, snaps: List[MultilayerNetwork]) -> None:
+    def __init__(self, snaps: list[MultilayerNetwork]) -> None:
         """
         Create a temporal network object.
 
@@ -88,7 +76,7 @@ class TemporalNetwork:
 
     @classmethod
     def from_nx_layers(
-        cls, network_list: List[nx.Graph], snap_ids: Optional[List[Any]] = None
+        cls, network_list: list[nx.Graph], snap_ids: list[Any] | None = None
     ) -> "TemporalNetwork":
         """
         Load a temporal network from a list of networks and their snapshot ids.
@@ -112,7 +100,7 @@ class TemporalNetwork:
             ]
         return cls(snaps)
 
-    def get_actors(self, shuffle: bool = False) -> List[MLNetworkActor]:
+    def get_actors(self, shuffle: bool = False) -> list[MLNetworkActor]:
         """
         Get actors that from the first snapshot of network.
 
@@ -122,7 +110,7 @@ class TemporalNetwork:
 
     def get_actors_from_snap(
         self, snapshot_id: int, shuffle: bool = False
-    ) -> List[MLNetworkActor]:
+    ) -> list[MLNetworkActor]:
         """
         Get actors that exist in the network at given snapshot.
 
@@ -188,4 +176,4 @@ class TemporalNetwork:
             MultilayerNetwork({"layer_1": cogsnet_snap_to_nxgraph(snap)})
             for snap in cogsnet_snaps
         ]
-        return cls(mln_snaps)  # TODO: add to docs!
+        return cls(mln_snaps)
